feat: remove popo module
