+++ src/linebot/handlers/message_handler.py
@@ -4,7 +4,7 @@ Message Event Handler
 from sqlalchemy.orm import Session
 from linebot.v3.webhooks import MessageEvent, TextMessageContent
 from linebot.v3.messaging import ReplyMessageRequest
-from src.popo.services import search_doctor
+from src.linebot.services import search_doctor
 from src.linebot.services import parse_search_criteria
 from src.linebot.message_templates.help_template import create_help_message
 from src.linebot.services import create_search_response
