+++ src/linebot/handlers/postback_handler.py
@@ -5,8 +5,8 @@ from urllib.parse import parse_qsl
 from sqlalchemy.orm import Session
 from linebot.v3.webhooks import PostbackEvent
 from linebot.v3.messaging import ReplyMessageRequest, TextMessage
-from src.popo.schemas import SearchCriteria, SearchType
-from src.popo.services import search_doctor
+from src.linebot.schemas import SearchCriteria, SearchType
+from src.linebot.services import search_doctor
 from src.linebot.dependencies import get_search_state
 from src.linebot.services import create_search_response
 from src.infra.logger import get_logger
