+++ src/main.py
@@ -6,7 +6,6 @@ from contextlib import asynccontextmanager
 from fastapi import FastAPI, APIRouter
 
 from src.linebot.router import router as linebot_router
-from src.popo.router import router as popo_router
 from src.linebot.dependencies import line_bot_api_wrapper
 from .config import settings
 
@@ -34,7 +33,6 @@ def create_app() -> FastAPI:
     )
 
     api_router = APIRouter(prefix="/api")
-    api_router.include_router(popo_router, prefix="/popo", tags=["popo"])
     api_router.include_router(linebot_router, prefix="/linebot", tags=["linebot"])
 
     _app.include_router(api_router)
